requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
aiosqlite==0.21.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
greenlet==3.2.1
h11==0.14.0
idna==3.6
pydantic==2.11.2
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.1
typing_extensions==4.13.1
uvicorn==0.34.0
python-multipart==0.0.9
Jinja2==3.1.3
aiofiles==23.2.1
supabase-py==2.3.4
httpx==0.27.0
postgrest==0.13.1
python-magic==0.4.27

